refacto: api calls
