+++ ui/components/core/src/api/inference/generation.ts
@@ -10,8 +10,6 @@ import { Annotation } from "../../lib/types";
 export async function conditional_generation_text_image(
   input: CondititionalGenerationTextImageInput,
 ): Promise<Annotation | null> {
-  let output: Annotation | null = null;
-
   try {
     const response = await fetch("/inference/tasks/conditional_generation/text-image", {
       headers: {
@@ -21,18 +19,20 @@ export async function conditional_generation_text_image(
       method: "POST",
       body: JSON.stringify(input),
     });
+
     if (!response.ok) {
-      console.log(
-        "api.conditional_generation_text_image -",
-        response.status,
-        response.statusText,
-        await response.text(),
-      );
-    } else {
-      output = (await response.json()) as Annotation;
+      // console.log(
+      //   "api.conditional_generation_text_image -",
+      //   response.status,
+      //   response.statusText,
+      //   await response.text(),
+      // );
+      return null;
     }
+
+    return (await response.json()) as Annotation;
   } catch (e) {
-    console.log("api.conditional_generation_text_image -", e);
+    // console.log("api.conditional_generation_text_image -", e);
+    return null;
   }
-  return output;
 }
