+++ ui/components/core/src/api/inference/index.ts
@@ -4,7 +4,8 @@ Author : pixano@cea.fr
 License: CECILL-C
 -------------------------------------*/
 
+export * from "./deleteModel";
+export * from "./generation";
+export * from "./instantiateModel";
 export * from "./isInferenceApiHealthy";
 export * from "./listModels";
-export * from "./instantiateModel";
-export * from "./generation";
