+++ ui/components/core/src/api/inference/listModels.ts
@@ -4,12 +4,14 @@ Author : pixano@cea.fr
 License: CECILL-C
 -------------------------------------*/
 
-import { type ModelList } from "../../lib/types";
+import { MultimodalImageNLPTask, type ModelList } from "../../lib/types";
 
-export async function listModels(task: string | null = null): Promise<ModelList[]> {
-  let model_list: ModelList[] = [];
-  let url = "/inference/models/list";
-  if (task) url += "?task=" + encodeURIComponent(task);
+export async function listModels(task: MultimodalImageNLPTask | null = null): Promise<ModelList[]> {
+  const url =
+    task === null
+      ? "/inference/models/list"
+      : `/inference/models/list?task=${encodeURIComponent(task)}`;
+
   try {
     const response = await fetch(url, {
       headers: {
@@ -18,13 +20,15 @@ export async function listModels(task: string | null = null): Promise<ModelList[
       },
       method: "GET",
     });
+
     if (!response.ok) {
       console.log("api.listModels -", response.status, response.statusText, await response.text());
-    } else {
-      model_list = (await response.json()) as ModelList[];
+      return [];
     }
+
+    return (await response.json()) as ModelList[];
   } catch (e) {
     console.log("api.listModels -", e);
+    return [];
   }
-  return model_list;
 }
